requirements.txt
SQLAlchemy==1.4.29
psycopg2==2.9.2
python-dotenv==0.19.2
python-telegram-bot==13.9